Allow other parsers
The code is already neatly designed such that users can almost supply
their own `Snapshot` class. This commit allows specifying the class in
the `ADExplorerSnapshot` constructor. For example, LDIF is another
format that contains the same data as an ADExplorer snapshot, and this
with this change one can easily pass their own class.